:truck: Moved DashBoard components into separate folder
UserList and PopUp moved into separate DashBoard folder to match file structure
